Separate all of the training routines for the constituency parser into their own module, rather than with the Trainer object
