+++ stanza/models/constituency/ensemble.py
@@ -33,9 +33,10 @@ from stanza.models.common import utils
 from stanza.models.common.foundation_cache import FoundationCache
 from stanza.models.constituency import retagging
 from stanza.models.constituency import tree_reader
+from stanza.models.constituency.parser_training import run_dev_set
 from stanza.models.constituency.state import MultiState
 from stanza.models.constituency.text_processing import parse_text, parse_dir
-from stanza.models.constituency.trainer import Trainer, run_dev_set
+from stanza.models.constituency.trainer import Trainer
 from stanza.models.constituency.utils import add_predict_output_args, postprocess_predict_output_args, retag_trees
 from stanza.resources.common import DEFAULT_MODEL_DIR
 from stanza.server.parser_eval import EvaluateParser, ParseResult, ScoredTree
