+++ stanza/models/constituency_parser.py
@@ -169,8 +169,8 @@ import stanza
 from stanza.models.common import constant
 from stanza.models.common import utils
 from stanza.models.common.peft_config import add_peft_args, resolve_peft_args
+from stanza.models.constituency import parser_training
 from stanza.models.constituency import retagging
-from stanza.models.constituency import trainer
 from stanza.models.constituency.lstm_model import ConstituencyComposition, SentenceBoundary, StackHistory
 from stanza.models.constituency.parse_transitions import TransitionScheme
 from stanza.models.constituency.text_processing import load_model_parse_text
@@ -849,13 +849,13 @@ def main(args=None):
     retag_pipeline = retagging.build_retag_pipeline(args)
 
     if args['mode'] == 'train':
-        trainer.train(args, model_load_file, retag_pipeline)
+        parser_training.train(args, model_load_file, retag_pipeline)
     elif args['mode'] == 'predict':
-        trainer.evaluate(args, model_load_file, retag_pipeline)
+        parser_training.evaluate(args, model_load_file, retag_pipeline)
     elif args['mode'] == 'parse_text':
         load_model_parse_text(args, model_load_file, retag_pipeline)
     elif args['mode'] == 'remove_optimizer':
-        trainer.remove_optimizer(args, args['save_name'], model_load_file)
+        parser_training.remove_optimizer(args, args['save_name'], model_load_file)
 
 if __name__ == '__main__':
     main()
